// src/types.ts
export type ActiveView =
  | 'home'
  | 'browse'
  | 'book-details'
  | 'sell'
  | 'orders'
  | 'wishlist'
  | 'requests'
  | 'notifications'
  | 'profile'
  | 'admin';